+++ common/src/main/java/net/sefacestudios/somemoreblocks/tags/SMBBiomeTags.java
@@ -28,6 +28,9 @@ public final class SMBBiomeTags {
   public static final TagKey<@NotNull Biome> GENERATES_SMALL_LILY_PADS = register("generates/small_lily_pads");
   public static final TagKey<@NotNull Biome> GENERATES_SNOW_GRASS = register("generates/snow_grass");
   public static final TagKey<@NotNull Biome> GENERATES_SPROUTS = register("generates/sprouts");
+  public static final TagKey<@NotNull Biome> GENERATES_DRY_SPROUTS = register("generates/dry_sprouts");
+  public static final TagKey<@NotNull Biome> GENERATES_PEBBLES = register("generates/pebbles");
+  public static final TagKey<@NotNull Biome> GENERATES_PEBBLES_STONY = register("generates/pebbles_stony");
   public static final TagKey<@NotNull Biome> GENERATES_SWAMP_VEGETATION = register("generates/swamp_vegetation");
   public static final TagKey<@NotNull Biome> GENERATES_WARPED_FUNGUS_COLONY = register("generates/warped_fungus_colony");
   public static final TagKey<@NotNull Biome> GENERATES_SNOW_BUSH = register("generates/snow_bush");