+++ fabric/src/main/java/net/sefacestudios/somemoreblocks/datagen/providers/assets/SMBModelProvider.java
@@ -10,8 +10,6 @@ import net.minecraft.client.data.models.ItemModelGenerators;
 import net.minecraft.client.data.models.ItemModelOutput;
 import net.minecraft.client.data.models.MultiVariant;
 import net.minecraft.client.data.models.blockstates.BlockModelDefinitionGenerator;
-import net.minecraft.client.data.models.blockstates.ConditionBuilder;
-import net.minecraft.client.data.models.blockstates.MultiPartGenerator;
 import net.minecraft.client.data.models.blockstates.MultiVariantGenerator;
 import net.minecraft.client.data.models.blockstates.PropertyDispatch;
 import net.minecraft.client.data.models.model.*;
@@ -324,6 +322,7 @@ public class SMBModelProvider extends FabricModelProvider {
     gen.createDoublePlantWithDefaultItem(SMBBlocks.RED_LARK_LARKSPUR.get(), BlockModelGenerators.PlantType.NOT_TINTED);
     gen.createPlantWithDefaultItem(SMBBlocks.SHORT_RED_LARK_LARKSPUR.get(), SMBBlocks.POTTED_SHORT_RED_LARK_LARKSPUR.get(), BlockModelGenerators.PlantType.NOT_TINTED);
     gen.createCrossBlock(SMBBlocks.SPROUTS.get(), BlockModelGenerators.PlantType.TINTED);
+    gen.createCrossBlockWithDefaultItem(SMBBlocks.DRY_SPROUTS.get(), BlockModelGenerators.PlantType.NOT_TINTED);
     gen.createItemWithGrassTint(SMBBlocks.SPROUTS.get());
     this.createDuckweed();
     this.createPebbles();
@@ -367,12 +366,12 @@ public class SMBModelProvider extends FabricModelProvider {
   private void createDuckweed() {
     Block block = SMBBlocks.DUCKWEED.get();
 
-    Identifier itemModel = this.blockModelGenerators.createFlatItemModelWithBlockTexture(SMBItems.DUCKWEED.get(), block, "_0");
+    Identifier itemModel = this.blockModelGenerators.createFlatItemModelWithBlockTexture(SMBItems.DUCKWEED.get(), block, "_1");
     this.blockModelGenerators.registerSimpleItemModel(block, itemModel);
 
     List<Weighted<Variant>> variants = new ArrayList<>();
 
-    for (int i = 0; i <= 2; i++) {
+    for (int i = 1; i <= 3; i++) {
       String suffix = "_" + i;
 
       TextureMapping textureMapping = TextureMapping.defaultTexture(block)
@@ -392,53 +391,25 @@ public class SMBModelProvider extends FabricModelProvider {
 
   private void createPebbles() {
     Block block = SMBBlocks.PEBBLES.get();
-    TextureMapping textureMapping = TextureMapping.defaultTexture(block)
-      .copyAndUpdate(TextureSlot.TEXTURE, TextureMapping.getBlockTexture(block));
 
-    MultiVariant pebbles1 = BlockModelGenerators.plainVariant(SMBModelTemplates.PEBBLES_1.create(block, textureMapping, this.modelOutput));
-    MultiVariant pebbles2 = BlockModelGenerators.plainVariant(SMBModelTemplates.PEBBLES_2.create(block, textureMapping, this.modelOutput));
-    MultiVariant pebbles3 = BlockModelGenerators.plainVariant(SMBModelTemplates.PEBBLES_3.create(block, textureMapping, this.modelOutput));
-    MultiVariant pebbles4 = BlockModelGenerators.plainVariant(SMBModelTemplates.PEBBLES_4.create(block, textureMapping, this.modelOutput));
+    MultiVariant models = new MultiVariant(WeightedList.of(
+      new Weighted<>(new Variant(ModelLocationUtils.getModelLocation(block, "_1")), 1),
+      new Weighted<>(new Variant(ModelLocationUtils.getModelLocation(block, "_2")), 1),
+      new Weighted<>(new Variant(ModelLocationUtils.getModelLocation(block, "_3")), 1)
+    ));
+
+    PropertyDispatch.C2<MultiVariant, Direction, Boolean> dispatch =
+      PropertyDispatch.initial(BlockStateProperties.HORIZONTAL_FACING, BlockStateProperties.WATERLOGGED);
+
+    for (boolean waterlogged : new boolean[]{false, true}) {
+      dispatch.select(Direction.NORTH, waterlogged, models);
+      dispatch.select(Direction.EAST, waterlogged, models.with(BlockModelGenerators.Y_ROT_90));
+      dispatch.select(Direction.SOUTH, waterlogged, models.with(BlockModelGenerators.Y_ROT_180));
+      dispatch.select(Direction.WEST, waterlogged, models.with(BlockModelGenerators.Y_ROT_270));
+    }
 
     this.blockModelGenerators.registerSimpleFlatItemModel(SMBItems.PEBBLES.get());
-
-    // Each template is one pebble piece (like flowerbed). Leaf-litter conditions are exclusive,
-    // so use additive segment_amount conditions instead.
-    MultiPartGenerator generator = MultiPartGenerator.multiPart(block);
-    this.addPebbleSegment(generator, Direction.NORTH, pebbles1, pebbles2, pebbles3, pebbles4, null);
-    this.addPebbleSegment(generator, Direction.EAST, pebbles1, pebbles2, pebbles3, pebbles4, BlockModelGenerators.Y_ROT_90);
-    this.addPebbleSegment(generator, Direction.SOUTH, pebbles1, pebbles2, pebbles3, pebbles4, BlockModelGenerators.Y_ROT_180);
-    this.addPebbleSegment(generator, Direction.WEST, pebbles1, pebbles2, pebbles3, pebbles4, BlockModelGenerators.Y_ROT_270);
-    this.blockStateOutput.accept(generator);
-  }
-
-  private void addPebbleSegment(
-    MultiPartGenerator generator,
-    Direction facing,
-    MultiVariant pebbles1,
-    MultiVariant pebbles2,
-    MultiVariant pebbles3,
-    MultiVariant pebbles4,
-    VariantMutator rotation
-  ) {
-    MultiVariant rotated1 = rotation == null ? pebbles1 : pebbles1.with(rotation);
-    MultiVariant rotated2 = rotation == null ? pebbles2 : pebbles2.with(rotation);
-    MultiVariant rotated3 = rotation == null ? pebbles3 : pebbles3.with(rotation);
-    MultiVariant rotated4 = rotation == null ? pebbles4 : pebbles4.with(rotation);
-
-    generator.with(new ConditionBuilder().term(BlockStateProperties.HORIZONTAL_FACING, facing), rotated1);
-    generator.with(
-      new ConditionBuilder().term(BlockStateProperties.HORIZONTAL_FACING, facing).term(BlockStateProperties.SEGMENT_AMOUNT, 2, 3, 4),
-      rotated2
-    );
-    generator.with(
-      new ConditionBuilder().term(BlockStateProperties.HORIZONTAL_FACING, facing).term(BlockStateProperties.SEGMENT_AMOUNT, 3, 4),
-      rotated3
-    );
-    generator.with(
-      new ConditionBuilder().term(BlockStateProperties.HORIZONTAL_FACING, facing).term(BlockStateProperties.SEGMENT_AMOUNT, 4),
-      rotated4
-    );
+    this.blockStateOutput.accept(MultiVariantGenerator.dispatch(block).with(dispatch));
   }
 
   /**